fix: update code coverage
